backend/app/blogs/views.py
from django.shortcuts import render
from rest_framework import authentication, permissions
from rest_framework.generics import ListAPIView, RetrieveAPIView, CreateAPIView
from rest_framework.response import Response
from rest_framework.views import APIView
from rest_framework.permissions import IsAuthenticated

from .models import Blog
from .serializers import BlogSerializer


class BlogListAPIView(ListAPIView):
    queryset = Blog.objects.all()
    serializer_class = BlogSerializer
    permission_classes = [IsAuthenticated]



class BlogCreateAPIView(CreateAPIView):
    serializer_class = BlogSerializer
    permission_classes = [IsAuthenticated]



class BlogDetailAPIView(RetrieveAPIView):
    queryset = Blog.objects.all()
    serializer_class = BlogSerializer
    permission_classes = [IsAuthenticated]

