backend/app/requirements.txt
django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
python-dotenv==1.1.0